+++ src/index.ts
@@ -37,7 +37,7 @@ import * as v2 from './v2';
 import {PassThrough, Duplex} from 'stream';
 import grpcGcpModule = require('grpc-gcp');
 import {ClusterUtils} from './utils/cluster';
-import { ClientSideMetricsController } from './client-side-metrics/metrics-controller';
+import { ClientSideMetricsConfigManager } from './client-side-metrics/metrics-config-manager';
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const streamEvents = require('stream-events');
@@ -435,7 +435,7 @@ export class Bigtable {
   static AppProfile: AppProfile;
   static Instance: Instance;
   static Cluster: Cluster;
-  metricsController: ClientSideMetricsController;
+  metricsConfigManager: ClientSideMetricsConfigManager;
 
   constructor(options: BigtableOptions = {}) {
     // Determine what scopes are needed.
@@ -535,9 +535,9 @@ export class Bigtable {
     this.shouldReplaceProjectIdToken = this.projectId === '{{projectId}}';
 
     if (options.metricsEnabled === false) {
-      this.metricsController = null;
+      this.metricsConfigManager = null;
     } else {
-      this.metricsController = ClientSideMetricsController(this.projectId, this.auth)
+      this.metricsConfigManager = new ClientSideMetricsConfigManager(this.projectId, this.auth)
     }
   }
