+++ src/tabular-api-surface.ts
@@ -339,8 +339,8 @@ Please use the format 'prezzy' or '${instance.name}/tables/prezzy'.`);
       }
       return originalEnd(chunk, encoding, cb);
     };
-    const metricsCollector = this.bigtable.metricsController
-      ? this.bigtable.metricsController.createOperation(
+    const metricsCollector = this.bigtable.metricsConfigManager
+      ? this.bigtable.metricsConfigManager.createOperation(
           this,
           MethodName.READ_ROWS,
           StreamingState.STREAMING,
